test(kubevirt): Add cluster creation test for storage driver ordering
Add TestCreateClusterKubeVirtStorageDriver to create_cluster_test.go
that:
- Creates a KubeVirt cluster with multiple storage class mappings
- Configures mappings equivalent to CLI flags:
  * --infra-storage-class-mapping=gp3-csi/gp3-csi
  * --infra-storage-class-mapping=a-gp3-csi/a-gp3-csi
  * --infra-volumesnapshot-class-mapping=csi-aws-vsc/csi-aws-vsc
  * --infra-volumesnapshot-class-mapping=a-csi-aws-vsc/a-csi-aws-vsc
- Validates driver-config ConfigMap has consistent alphabetical ordering
- Checks ConfigMap content doesn't flap across reconciliations

This test validates OCPBUGS-61245 fix by creating a full cluster with
storage driver configuration and verifying the ConfigMap ordering
remains
stable, complementing the standalone kubevirt_storage_driver_test.go.
